add public api fixes
This PR fixes issues with the Public api generation related to the Metrics, snowflake and Delta Lake api.

There still persists an issue with the Network service and an import missing from the original proto file import "google/protobuf/wrappers.proto"; which results in the import in the public api being removed.
This should be fixed by the responsible team.