. Fix missing field in `GoogleAds` source endpoint
`Credentials` was a public message with public fields, but its only
usage `credentials = 7` was not marked as public. As a result, it is
missing from the interface and from the public API, rendering GoogleAds
source endpoint unusable.

Fix this by adding a proper modifier of API visibility to `Credentials`
and adding a proper translation key for it.